+++ flatlaf-core/src/main/java/com/formdev/flatlaf/IntelliJTheme.java
@@ -16,7 +16,6 @@
 
 package com.formdev.flatlaf;
 
-import java.awt.Color;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -25,20 +24,16 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 import java.util.Set;
-import java.util.Map.Entry;
-import javax.swing.UIDefaults;
-import javax.swing.plaf.ColorUIResource;
 import com.formdev.flatlaf.json.Json;
 import com.formdev.flatlaf.json.ParseException;
-import com.formdev.flatlaf.util.ColorFunctions;
 import com.formdev.flatlaf.util.LoggingFacade;
 import com.formdev.flatlaf.util.StringUtils;
 import com.formdev.flatlaf.util.SystemInfo;
@@ -63,13 +58,11 @@ public class IntelliJTheme
 	public final boolean dark;
 	public final String author;
 
-	private final boolean isMaterialUILite;
+	private Map<String, String> jsonColors;
+	private Map<String, Object> jsonUI;
+	private Map<String, Object> jsonIcons;
 
-	private Map<String, String> colors;
-	private Map<String, Object> ui;
-	private Map<String, Object> icons;
-
-	private Map<String, ColorUIResource> namedColors = Collections.emptyMap();
+	private Map<String, String> namedColors = Collections.emptyMap();
 
 	/**
 	 * Loads a IntelliJ .theme.json file from the given input stream,
@@ -84,7 +77,7 @@ public class IntelliJTheme
 		try {
 		    return FlatLaf.setup( createLaf( in ) );
 		} catch( Exception ex ) {
-			LoggingFacade.INSTANCE.logSevere(  "FlatLaf: Failed to load IntelliJ theme", ex );
+			LoggingFacade.INSTANCE.logSevere( "FlatLaf: Failed to load IntelliJ theme", ex );
 		    return false;
 		}
 	}
@@ -138,94 +131,90 @@ public class IntelliJTheme
 	    dark = Boolean.parseBoolean( (String) json.get( "dark" ) );
 	    author = (String) json.get( "author" );
 
-		isMaterialUILite = author.equals( "Mallowigi" );
-
-	    colors = (Map<String, String>) json.get( "colors" );
-	    ui = (Map<String, Object>) json.get( "ui" );
-	    icons = (Map<String, Object>) json.get( "icons" );
+	    jsonColors = (Map<String, String>) json.get( "colors" );
+	    jsonUI = (Map<String, Object>) json.get( "ui" );
+	    jsonIcons = (Map<String, Object>) json.get( "icons" );
 	}
 
-	private void applyProperties( UIDefaults defaults ) {
-		if( ui == null )
+	private void applyProperties( Properties properties ) {
+		if( jsonUI == null )
 			return;
 
-		defaults.put( "Component.isIntelliJTheme", true );
+		put( properties, "Component.isIntelliJTheme", "true" );
 
 		// enable button shadows
-		defaults.put( "Button.paintShadow", true );
-		defaults.put( "Button.shadowWidth", dark ? 2 : 1 );
+		put( properties, "Button.paintShadow", "true" );
+		put( properties, "Button.shadowWidth", dark ? "2" : "1" );
 
-		Map<Object, Object> themeSpecificDefaults = removeThemeSpecificDefaults( defaults );
+		Map<String, String> themeSpecificProps = removeThemeSpecificProps( properties );
+		Set<String> jsonUIKeys = new HashSet<>();
 
-		loadNamedColors( defaults );
+		// Json node "colors"
+		loadNamedColors( properties, jsonUIKeys );
 
-		// convert Json "ui" structure to UI defaults
-		ArrayList<Object> defaultsKeysCache = new ArrayList<>();
-		Set<String> uiKeys = new HashSet<>();
-		for( Map.Entry<String, Object> e : ui.entrySet() )
-			apply( e.getKey(), e.getValue(), defaults, defaultsKeysCache, uiKeys );
+		// convert Json "ui" structure to UI properties
+		for( Map.Entry<String, Object> e : jsonUI.entrySet() )
+			apply( e.getKey(), e.getValue(), properties, jsonUIKeys );
 
-		applyColorPalette( defaults );
-		applyCheckBoxColors( defaults );
+		// set FlatLaf variables
+		copyIfSetInJson( properties, jsonUIKeys, "@background", "Panel.background", "*.background" );
+		copyIfSetInJson( properties, jsonUIKeys, "@foreground", "CheckBox.foreground", "*.foreground" );
+		copyIfSetInJson( properties, jsonUIKeys, "@accentBaseColor",
+			"ColorPalette.accent", // Material UI Lite, Hiberbee
+			"ColorPalette.accentColor", // Dracula, One Dark
+			"ProgressBar.foreground",
+			"*.selectionBackground" );
+		copyIfSetInJson( properties, jsonUIKeys, "@accentUnderlineColor", "*.underlineColor", "TabbedPane.underlineColor" );
+		copyIfSetInJson( properties, jsonUIKeys, "@selectionBackground", "*.selectionBackground" );
+		copyIfSetInJson( properties, jsonUIKeys, "@selectionForeground", "*.selectionForeground" );
+		copyIfSetInJson( properties, jsonUIKeys, "@selectionInactiveBackground", "*.selectionInactiveBackground" );
+		copyIfSetInJson( properties, jsonUIKeys, "@selectionInactiveForeground", "*.selectionInactiveForeground" );
+
+		// Json node "icons/ColorPalette"
+		applyIconsColorPalette( properties );
+
+		// apply "CheckBox.icon." colors
+		applyCheckBoxColors( properties );
 
 		// copy values
 		for( Map.Entry<String, String> e : uiKeyCopying.entrySet() ) {
-			Object value = defaults.get( e.getValue() );
+			Object value = properties.get( e.getValue() );
 			if( value != null )
-				defaults.put( e.getKey(), value );
+				put( properties, e.getKey(), value );
 		}
 
 		// IDEA does not paint button background if disabled, but FlatLaf does
-		Object panelBackground = defaults.get( "Panel.background" );
-		defaults.put( "Button.disabledBackground", panelBackground );
-		defaults.put( "ToggleButton.disabledBackground", panelBackground );
+		put( properties, "Button.disabledBackground", "@disabledBackground" );
+		put( properties, "ToggleButton.disabledBackground", "@disabledBackground" );
 
-		// fix Button borders
-		copyIfNotSet( defaults, "Button.focusedBorderColor", "Component.focusedBorderColor", uiKeys );
-		defaults.put( "Button.hoverBorderColor", defaults.get( "Button.focusedBorderColor" ) );
-		defaults.put( "HelpButton.hoverBorderColor", defaults.get( "Button.focusedBorderColor" ) );
-
-		// IDEA uses an SVG icon for the help button, but paints the background with Button.startBackground and Button.endBackground
-		Object helpButtonBackground = defaults.get( "Button.startBackground" );
-		Object helpButtonBorderColor = defaults.get( "Button.startBorderColor" );
-		if( helpButtonBackground == null )
-			helpButtonBackground = defaults.get( "Button.background" );
-		if( helpButtonBorderColor == null )
-			helpButtonBorderColor = defaults.get( "Button.borderColor" );
-		defaults.put( "HelpButton.background", helpButtonBackground );
-		defaults.put( "HelpButton.borderColor", helpButtonBorderColor );
-		defaults.put( "HelpButton.disabledBackground", panelBackground );
-		defaults.put( "HelpButton.disabledBorderColor", defaults.get( "Button.disabledBorderColor" ) );
-		defaults.put( "HelpButton.focusedBorderColor", defaults.get( "Button.focusedBorderColor" ) );
-		defaults.put( "HelpButton.focusedBackground", defaults.get( "Button.focusedBackground" ) );
+		// fix Button
+		fixStartEnd( properties, jsonUIKeys, "Button.startBackground", "Button.endBackground", "Button.background" );
+		fixStartEnd( properties, jsonUIKeys, "Button.startBorderColor", "Button.endBorderColor", "Button.borderColor" );
+		fixStartEnd( properties, jsonUIKeys, "Button.default.startBackground", "Button.default.endBackground", "Button.default.background" );
+		fixStartEnd( properties, jsonUIKeys, "Button.default.startBorderColor", "Button.default.endBorderColor", "Button.default.borderColor" );
 
 		// IDEA uses TextField.background for editable ComboBox and Spinner
-		Object textFieldBackground = get( defaults, themeSpecificDefaults, "TextField.background" );
-		defaults.put( "ComboBox.editableBackground", textFieldBackground );
-		defaults.put( "Spinner.background", textFieldBackground );
-
-		// Spinner arrow button always has same colors as ComboBox arrow button
-		defaults.put( "Spinner.buttonBackground", defaults.get( "ComboBox.buttonEditableBackground" ) );
-		defaults.put( "Spinner.buttonArrowColor", defaults.get( "ComboBox.buttonArrowColor" ) );
-		defaults.put( "Spinner.buttonDisabledArrowColor", defaults.get( "ComboBox.buttonDisabledArrowColor" ) );
+		Object textFieldBackground = get( properties, themeSpecificProps, "TextField.background" );
+		put( properties, "ComboBox.editableBackground", textFieldBackground );
+		put( properties, "Spinner.background", textFieldBackground );
 
 		// some themes specify colors for TextField.background, but forget to specify it for other components
 		// (probably because those components are not used in IntelliJ IDEA)
-		putAll( defaults, textFieldBackground,
+		putAll( properties, textFieldBackground,
 			"EditorPane.background",
 			"FormattedTextField.background",
 			"PasswordField.background",
 			"TextArea.background",
 			"TextPane.background"
 		);
-		putAll( defaults, get( defaults, themeSpecificDefaults, "TextField.selectionBackground" ),
+		putAll( properties, get( properties, themeSpecificProps, "TextField.selectionBackground" ),
 			"EditorPane.selectionBackground",
 			"FormattedTextField.selectionBackground",
 			"PasswordField.selectionBackground",
 			"TextArea.selectionBackground",
 			"TextPane.selectionBackground"
 		);
-		putAll( defaults, get( defaults, themeSpecificDefaults, "TextField.selectionForeground" ),
+		putAll( properties, get( properties, themeSpecificProps, "TextField.selectionForeground" ),
 			"EditorPane.selectionForeground",
 			"FormattedTextField.selectionForeground",
 			"PasswordField.selectionForeground",
@@ -235,7 +224,7 @@ public class IntelliJTheme
 
 		// fix disabled and not-editable backgrounds for text components, combobox and spinner
 		// (IntelliJ IDEA does not use those colors; instead it used background color of parent)
-		putAll( defaults, panelBackground,
+		putAll( properties, "@disabledBackground",
 			"ComboBox.disabledBackground",
 			"EditorPane.disabledBackground", "EditorPane.inactiveBackground",
 			"FormattedTextField.disabledBackground", "FormattedTextField.inactiveBackground",
@@ -246,132 +235,148 @@ public class IntelliJTheme
 			"TextPane.disabledBackground", "TextPane.inactiveBackground"
 		);
 
-		// fix ToggleButton
-		if( !uiKeys.contains( "ToggleButton.startBackground" ) && !uiKeys.contains( "*.startBackground" ) )
-			defaults.put( "ToggleButton.startBackground", defaults.get( "Button.startBackground" ) );
-		if( !uiKeys.contains( "ToggleButton.endBackground" ) && !uiKeys.contains( "*.endBackground" ) )
-			defaults.put( "ToggleButton.endBackground", defaults.get( "Button.endBackground" ) );
-		if( !uiKeys.contains( "ToggleButton.foreground" ) && uiKeys.contains( "Button.foreground" ) )
-			defaults.put( "ToggleButton.foreground", defaults.get( "Button.foreground" ) );
-
 		// fix DesktopPane background (use Panel.background and make it 5% darker/lighter)
-		Color desktopBackgroundBase = defaults.getColor( "Panel.background" );
-		Color desktopBackground = ColorFunctions.applyFunctions( desktopBackgroundBase,
-			new ColorFunctions.HSLIncreaseDecrease( 2, dark, 5, false, true ) );
-		defaults.put( "Desktop.background", new ColorUIResource( desktopBackground ) );
-
-		// fix List and Table background colors in Material UI Lite themes
-		if( isMaterialUILite ) {
-			defaults.put( "List.background", defaults.get( "Tree.background" ) );
-			defaults.put( "Table.background", defaults.get( "Tree.background" ) );
-		}
+		put( properties, "Desktop.background", dark ? "lighten($Panel.background,5%)" : "darken($Panel.background,5%)" );
 
 		// limit tree row height
-		int rowHeight = defaults.getInt( "Tree.rowHeight" );
+		String rowHeightStr = (String) properties.get( "Tree.rowHeight" );
+		int rowHeight = (rowHeightStr != null) ? Integer.parseInt( rowHeightStr ) : 0;
 		if( rowHeight > 22 )
-			defaults.put( "Tree.rowHeight", 22 );
+			put( properties, "Tree.rowHeight", "22" );
 
-		// get (and remove) theme specific wildcard replacements, which override all other defaults that end with same suffix
-		HashMap<String, Object> wildcards = new HashMap<>();
-		Iterator<Entry<Object, Object>> it = themeSpecificDefaults.entrySet().iterator();
+		// get (and remove) theme specific wildcard replacements, which override all other properties that end with same suffix
+		HashMap<String, String> wildcardProps = new HashMap<>();
+		Iterator<Map.Entry<String, String>> it = themeSpecificProps.entrySet().iterator();
 		while( it.hasNext() ) {
-			Entry<Object, Object> e = it.next();
-			String key = (String) e.getKey();
+			Map.Entry<String, String> e = it.next();
+			String key = e.getKey();
 			if( key.startsWith( "*." ) ) {
-				wildcards.put( key.substring( "*.".length() ), e.getValue() );
+				wildcardProps.put( key, e.getValue() );
 				it.remove();
 			}
 		}
 
-		// override UI defaults with theme specific wildcard replacements
-		if( !wildcards.isEmpty() ) {
-			for( Object key : defaults.keySet().toArray() ) {
-				int dot;
-				if( !(key instanceof String) ||
-					(dot = ((String)key).lastIndexOf( '.' )) < 0 )
-				  continue;
-
-				String wildcardKey = ((String)key).substring( dot + 1 );
-				Object wildcardValue = wildcards.get( wildcardKey );
-				if( wildcardValue != null )
-					defaults.put( key, wildcardValue );
-			}
+		// override properties with theme specific wildcard replacements
+		if( !wildcardProps.isEmpty() ) {
+			for( Map.Entry<String, String> e : wildcardProps.entrySet() )
+				applyWildcard( properties, e.getKey(), e.getValue() );
 		}
 
-		// apply theme specific UI defaults at the end to allow overwriting
-		for( Map.Entry<Object, Object> e : themeSpecificDefaults.entrySet() ) {
-			Object key = e.getKey();
-			Object value = e.getValue();
+		// apply theme specific properties at the end to allow overwriting
+		for( Map.Entry<String, String> e : themeSpecificProps.entrySet() ) {
+			String key = e.getKey();
+			String value = e.getValue();
 
 			// append styles to existing styles
-			if( key instanceof String && ((String)key).startsWith( "[style]" ) ) {
-				Object oldValue = defaults.get( key );
+			if( key.startsWith( "[style]" ) ) {
+				String oldValue = (String) properties.get( key );
 				if( oldValue != null )
 					value = oldValue + "; " + value;
 			}
 
-			defaults.put( key, value );
+			put( properties, key, value );
 		}
 
 		// let Java release memory
-		colors = null;
-		ui = null;
-		icons = null;
+		jsonColors = null;
+		jsonUI = null;
+		jsonIcons = null;
 	}
 
-	private Object get( UIDefaults defaults, Map<Object, Object> themeSpecificDefaults, String key ) {
-		return themeSpecificDefaults.getOrDefault( key, defaults.get( key ) );
+	private String get( Properties properties, Map<String, String> themeSpecificProps, String key ) {
+		return themeSpecificProps.getOrDefault( key, (String) properties.get( key ) );
 	}
 
-	private void putAll( UIDefaults defaults, Object value, String... keys ) {
+	private void put( Properties properties, Object key, Object value ) {
+		if( value != null )
+			properties.put( key, value );
+		else
+			properties.remove( key );
+	}
+
+	private void putAll( Properties properties, Object value, String... keys ) {
 		for( String key : keys )
-			defaults.put( key, value );
+			put( properties, key, value );
 	}
 
-	private Map<Object, Object> removeThemeSpecificDefaults( UIDefaults defaults ) {
-		// search for theme specific UI defaults keys
+	private void copyIfSetInJson( Properties properties, Set<String> jsonUIKeys, String destKey, String... srcKeys ) {
+		for( String srcKey : srcKeys ) {
+			if( jsonUIKeys.contains( srcKey ) ) {
+				Object value = properties.get( srcKey );
+				if( value != null ) {
+					put( properties, destKey, value );
+					break;
+				}
+			}
+		}
+	}
+
+	private void fixStartEnd( Properties properties, Set<String> jsonUIKeys, String startKey, String endKey, String key ) {
+		if( jsonUIKeys.contains( startKey ) && jsonUIKeys.contains( endKey ) )
+			put( properties, key, "$" + startKey );
+	}
+
+	private Map<String, String> removeThemeSpecificProps( Properties properties ) {
+		// search for theme specific properties keys
 		ArrayList<String> themeSpecificKeys = new ArrayList<>();
-		for( Object key : defaults.keySet() ) {
-			if( key instanceof String && ((String)key).startsWith( "{" ) )
+		for( Object key : properties.keySet() ) {
+			if( ((String)key).startsWith( "{" ) )
 				themeSpecificKeys.add( (String) key );
 		}
 
-		// remove theme specific UI defaults and remember only those for current theme
-		Map<Object, Object> themeSpecificDefaults = new HashMap<>();
+		// special prefixes (priority from highest to lowest)
 		String currentThemePrefix = '{' + name.replace( ' ', '_' ) + '}';
 		String currentThemeAndAuthorPrefix = '{' + name.replace( ' ', '_' ) + "---" + author.replace( ' ', '_' ) + '}';
 		String currentAuthorPrefix = "{author-" + author.replace( ' ', '_' ) + '}';
+		String lightOrDarkPrefix = dark ? "{*-dark}" : "{*-light}";
 		String allThemesPrefix = "{*}";
-		String[] prefixes = { currentThemePrefix, currentThemeAndAuthorPrefix, currentAuthorPrefix, allThemesPrefix };
+		String[] prefixes = { currentThemePrefix, currentThemeAndAuthorPrefix, currentAuthorPrefix, lightOrDarkPrefix, allThemesPrefix };
+
+		// collect values for special prefixes in its own maps
+		@SuppressWarnings( "unchecked" )
+		Map<String, String>[] maps = new Map[prefixes.length];
+		for( int i = 0; i < maps.length; i++ )
+			maps[i] = new HashMap<>();
+
+		// remove theme specific properties and remember only those for current theme
 		for( String key : themeSpecificKeys ) {
-			Object value = defaults.remove( key );
-			for( String prefix : prefixes ) {
+			String value = (String) properties.remove( key );
+			for( int i = 0; i < prefixes.length; i++ ) {
+				String prefix = prefixes[i];
 				if( key.startsWith( prefix ) ) {
-					themeSpecificDefaults.put( key.substring( prefix.length() ), value );
+					maps[i].put( key.substring( prefix.length() ), value );
 					break;
 				}
 			}
 		}
 
-		return themeSpecificDefaults;
+		// copy values into single map (from lowest to highest priority)
+		Map<String, String> themeSpecificProps = new HashMap<>();
+		for( int i = maps.length - 1; i >= 0; i-- )
+			themeSpecificProps.putAll( maps[i] );
+		return themeSpecificProps;
 	}
 
 	/**
 	 * http://www.jetbrains.org/intellij/sdk/docs/reference_guide/ui_themes/themes_customize.html#defining-named-colors
 	 */
-	private void loadNamedColors( UIDefaults defaults ) {
-		if( colors == null )
+	private void loadNamedColors( Properties properties, Set<String> jsonUIKeys ) {
+		if( jsonColors == null )
 			return;
 
 		namedColors = new HashMap<>();
 
-		for( Map.Entry<String, String> e : colors.entrySet() ) {
+		for( Map.Entry<String, String> e : jsonColors.entrySet() ) {
 			String value = e.getValue();
-			ColorUIResource color = parseColor( value );
-			if( color != null ) {
+			if( canParseColor( value ) ) {
 				String key = e.getKey();
-				namedColors.put( key, color );
-				defaults.put( "ColorPalette." + key, color );
+				namedColors.put( key, value );
+
+				String uiKey = "ColorPalette." + key;
+				put( properties, uiKey, value );
+
+				// this is only necessary for copyIfSetInJson() (used for accent color)
+				jsonUIKeys.add( uiKey );
 			}
 		}
 	}
@@ -380,7 +385,7 @@ public class IntelliJTheme
 	 * http://www.jetbrains.org/intellij/sdk/docs/reference_guide/ui_themes/themes_customize.html#custom-ui-control-colors
 	 */
 	@SuppressWarnings( "unchecked" )
-	private void apply( String key, Object value, UIDefaults defaults, ArrayList<Object> defaultsKeysCache, Set<String> uiKeys ) {
+	private void apply( String key, Object value, Properties properties, Set<String> jsonUIKeys ) {
 		if( value instanceof Map ) {
 			Map<String, Object> map = (Map<String, Object>)value;
 			if( map.containsKey( "os.default" ) || map.containsKey( "os.windows" ) || map.containsKey( "os.mac" ) || map.containsKey( "os.linux" ) ) {
@@ -388,12 +393,12 @@ public class IntelliJTheme
 					: SystemInfo.isMacOS ? "os.mac"
 					: SystemInfo.isLinux ? "os.linux" : null;
 				if( osKey != null && map.containsKey( osKey ) )
-					apply( key, map.get( osKey ), defaults, defaultsKeysCache, uiKeys );
+					apply( key, map.get( osKey ), properties, jsonUIKeys );
 				else if( map.containsKey( "os.default" ) )
-					apply( key, map.get( "os.default" ), defaults, defaultsKeysCache, uiKeys );
+					apply( key, map.get( "os.default" ), properties, jsonUIKeys );
 			} else {
 				for( Map.Entry<String, Object> e : map.entrySet() )
-					apply( key + '.' + e.getKey(), e.getValue(), defaults, defaultsKeysCache, uiKeys );
+					apply( key + '.' + e.getKey(), e.getValue(), properties, jsonUIKeys );
 			}
 		} else {
 			if( "".equals( value ) )
@@ -418,15 +423,15 @@ public class IntelliJTheme
 			if( dot > 0 && uiKeyExcludes.contains( key.substring( 0, dot + 1 ) ) )
 				return;
 
-			if( uiKeyDoNotOverride.contains( key ) && uiKeys.contains( key ) )
+			if( uiKeyDoNotOverride.contains( key ) && jsonUIKeys.contains( key ) )
 				return;
 
-			uiKeys.add( key );
+			jsonUIKeys.add( key );
 
 			String valueStr = value.toString();
 
 			// map named colors
-			Object uiValue = namedColors.get( valueStr );
+			String uiValue = namedColors.get( valueStr );
 
 			// parse value
 			if( uiValue == null ) {
@@ -445,45 +450,62 @@ public class IntelliJTheme
 
 				// parse value
 				try {
-					uiValue = UIDefaultsLoader.parseValue( key, valueStr, null );
+					UIDefaultsLoader.parseValue( key, valueStr, null );
+					uiValue = valueStr;
 				} catch( RuntimeException ex ) {
-					UIDefaultsLoader.logParseError( key, valueStr, ex, false );
+					UIDefaultsLoader.logParseError( key, valueStr, ex, true );
 					return; // ignore invalid value
 				}
 			}
 
-			if( key.startsWith( "*." ) ) {
-				// wildcard
-				String tail = key.substring( 1 );
+			// wildcards
+			if( applyWildcard( properties, key, uiValue ) )
+				return;
 
-				// because we can not iterate over the UI defaults keys while
-				// modifying UI defaults in the same loop, we have to copy the keys
-				if( defaultsKeysCache.size() != defaults.size() ) {
-					defaultsKeysCache.clear();
-					Enumeration<Object> e = defaults.keys();
-					while( e.hasMoreElements() )
-						defaultsKeysCache.add( e.nextElement() );
-				}
-
-				// replace all values in UI defaults that match the wildcard key
-				for( Object k : defaultsKeysCache ) {
-					if( k.equals( "Desktop.background" ) ||
-						k.equals( "DesktopIcon.background" ) ||
-						k.equals( "TabbedPane.focusColor" ) )
-					  continue;
-
-					if( k instanceof String ) {
-						// support replacing of mapped keys
-						// (e.g. set ComboBox.buttonEditableBackground to *.background
-						// because it is mapped from ComboBox.ArrowButton.background)
-						String km = uiKeyInverseMapping.getOrDefault( k, (String) k );
-						if( km.endsWith( tail ) && !((String)k).startsWith( "CheckBox.icon." ) )
-							defaults.put( k, uiValue );
-					}
-				}
-			} else
-				defaults.put( key, uiValue );
+			put( properties, key, uiValue );
 		}
+	}
+
+	private boolean applyWildcard( Properties properties, String key, String value ) {
+		if( !key.startsWith( "*." ) )
+			return false;
+
+		String tail = key.substring( 1 );
+
+		// because we can not iterate over the properties keys while
+		// modifying properties in the same loop, we have to copy the keys
+		String[] keys = properties.keySet().toArray( new String[properties.size()] );
+
+		// replace all values in properties that match the wildcard key
+		for( String k : keys ) {
+			if( k.startsWith( "*" ) ||
+				k.startsWith( "@" ) ||
+				k.startsWith( "HelpButton." ) ||
+				k.startsWith( "JX" ) ||
+				k.startsWith( "Jide" ) ||
+				k.startsWith( "ProgressBar.selection" ) ||
+				k.startsWith( "TitlePane." ) ||
+				k.startsWith( "ToggleButton.tab." ) ||
+				k.equals( "Desktop.background" ) ||
+				k.equals( "DesktopIcon.background" ) ||
+				k.equals( "TabbedPane.focusColor" ) ||
+				k.endsWith( ".hoverBackground" ) ||
+				k.endsWith( ".pressedBackground" ) )
+			  continue;
+
+			// support replacing of mapped keys
+			// (e.g. set ComboBox.buttonEditableBackground to *.background
+			// because it is mapped from ComboBox.ArrowButton.background)
+			String km = uiKeyInverseMapping.getOrDefault( k, k );
+			if( km.endsWith( tail ) && !k.startsWith( "CheckBox.icon." ) )
+				put( properties, k, value );
+		}
+
+		// Note: also add wildcards to properties and let UIDefaultsLoader
+		//       process it on BasicLookAndFeel UI defaults
+		put( properties, key, value );
+
+		return true;
 	}
 
 	private String fixColorIfValid( String newColorStr, String colorStr ) {
@@ -497,11 +519,11 @@ public class IntelliJTheme
 		}
 	}
 
-	private void applyColorPalette( UIDefaults defaults ) {
-		if( icons == null )
+	private void applyIconsColorPalette( Properties properties ) {
+		if( jsonIcons == null )
 			return;
 
-		Object palette = icons.get( "ColorPalette" );
+		Object palette = jsonIcons.get( "ColorPalette" );
 		if( !(palette instanceof Map) )
 			return;
 
@@ -510,44 +532,48 @@ public class IntelliJTheme
 		for( Map.Entry<String, Object> e : colorPalette.entrySet() ) {
 			String key = e.getKey();
 			Object value = e.getValue();
-			if( key.startsWith( "Checkbox." ) || !(value instanceof String) )
+			if( key.startsWith( "Checkbox." ) || key.startsWith( "#" ) || !(value instanceof String) )
 				continue;
 
 			if( dark )
 				key = StringUtils.removeTrailing( key, ".Dark" );
 
-			ColorUIResource color = toColor( (String) value );
+			String color = toColor( (String) value );
 			if( color != null )
-				defaults.put( key, color );
+				put( properties, key, color );
 		}
 	}
 
-	private ColorUIResource toColor( String value ) {
+	private String toColor( String value ) {
+		if( value.startsWith( "##" ) )
+			value = fixColorIfValid( value.substring( 1 ), value );
+
 		// map named colors
-		ColorUIResource color = namedColors.get( value );
+		String color = namedColors.get( value );
 
 		// parse color
-		return (color != null) ? color : parseColor( value );
+		return (color != null) ? color : (canParseColor( value ) ? value : null);
 	}
 
-	private ColorUIResource parseColor( String value ) {
+	private boolean canParseColor( String value ) {
 		try {
-			return UIDefaultsLoader.parseColor( value );
+			return UIDefaultsLoader.parseColor( value ) != null;
 		} catch( IllegalArgumentException ex ) {
-			return null;
+			LoggingFacade.INSTANCE.logSevere( "FlatLaf: Failed to parse color: '" + value + '\'', ex );
+			return false;
 		}
 	}
 
 	/**
 	 * Because IDEA uses SVGs for check boxes and radio buttons, the colors for
 	 * these two components are specified in "icons > ColorPalette".
-	 * FlatLaf uses vector icons and expects colors for the two components in UI defaults.
+	 * FlatLaf uses vector icons and expects colors for the two components in properties.
 	 */
-	private void applyCheckBoxColors( UIDefaults defaults ) {
-		if( icons == null )
+	private void applyCheckBoxColors( Properties properties ) {
+		if( jsonIcons == null )
 			return;
 
-		Object palette = icons.get( "ColorPalette" );
+		Object palette = jsonIcons.get( "ColorPalette" );
 		if( !(palette instanceof Map) )
 			return;
 
@@ -569,9 +595,9 @@ public class IntelliJTheme
 				if( !dark && newKey.startsWith( checkBoxIconPrefix ) )
 					newKey = "CheckBox.icon[filled].".concat( newKey.substring( checkBoxIconPrefix.length() ) );
 
-				ColorUIResource color = toColor( (String) value );
+				String color = toColor( (String) value );
 				if( color != null ) {
-					defaults.put( newKey, color );
+					put( properties, newKey, color );
 
 					String key2 = checkboxDuplicateColors.get( key + ".Dark");
 					if( key2 != null ) {
@@ -592,7 +618,7 @@ public class IntelliJTheme
 
 						String newKey2 = checkboxKeyMapping.get( key2 );
 						if( newKey2 != null )
-							defaults.put( newKey2, color );
+							put( properties, newKey2, color );
 					}
 				}
 
@@ -603,13 +629,13 @@ public class IntelliJTheme
 		// update hover, pressed and focused colors
 		if( checkboxModified ) {
 			// for non-filled checkbox/radiobutton used in dark themes
-			defaults.remove( "CheckBox.icon.focusWidth" );
-			defaults.put( "CheckBox.icon.hoverBorderColor", defaults.get( "CheckBox.icon.focusedBorderColor" ) );
+			properties.remove( "CheckBox.icon.focusWidth" );
+			put( properties, "CheckBox.icon.hoverBorderColor", properties.get( "CheckBox.icon.focusedBorderColor" ) );
 
 			// for filled checkbox/radiobutton used in light themes
-			defaults.remove( "CheckBox.icon[filled].focusWidth" );
-			defaults.put( "CheckBox.icon[filled].hoverBorderColor", defaults.get( "CheckBox.icon[filled].focusedBorderColor" ) );
-			defaults.put( "CheckBox.icon[filled].focusedSelectedBackground", defaults.get( "CheckBox.icon[filled].selectedBackground" ) );
+			properties.remove( "CheckBox.icon[filled].focusWidth" );
+			put( properties, "CheckBox.icon[filled].hoverBorderColor", properties.get( "CheckBox.icon[filled].focusedBorderColor" ) );
+			put( properties, "CheckBox.icon[filled].focusedSelectedBackground", properties.get( "CheckBox.icon[filled].selectedBackground" ) );
 
 			if( dark ) {
 				// IDEA Darcula checkBoxFocused.svg, checkBoxSelectedFocused.svg,
@@ -623,19 +649,12 @@ public class IntelliJTheme
 					"CheckBox.icon[filled].focusedSelectedBorderColor",
 				};
 				for( String key : focusedBorderColorKeys ) {
-					Color color = defaults.getColor( key );
-					if( color != null ) {
-						defaults.put( key, new ColorUIResource( new Color(
-							(color.getRGB() & 0xffffff) | 0xa6000000, true ) ) );
-					}
+					Object color = properties.get( key );
+					if( color != null )
+						put( properties, key, "fade(" + color + ", 65%)" );
 				}
 			}
 		}
-	}
-
-	private void copyIfNotSet( UIDefaults defaults, String destKey, String srcKey, Set<String> uiKeys ) {
-		if( !uiKeys.contains( destKey ) )
-			defaults.put( destKey, defaults.get( srcKey ) );
 	}
 
 	private static final Set<String> uiKeyExcludes;
@@ -653,26 +672,26 @@ public class IntelliJTheme
 		uiKeyExcludes = new HashSet<>( Arrays.asList(
 			"ActionButton.", "ActionToolbar.", "ActionsList.", "AppInspector.", "AssignedMnemonic.", "Autocomplete.",
 			"AvailableMnemonic.",
-			"Badge.", "BigSpinner.", "Bookmark.", "BookmarkIcon.", "BookmarkMnemonicAssigned.", "BookmarkMnemonicAvailable.",
+			"Badge.", "Banner.", "BigSpinner.", "Bookmark.", "BookmarkIcon.", "BookmarkMnemonicAssigned.", "BookmarkMnemonicAvailable.",
 			"BookmarkMnemonicCurrent.", "BookmarkMnemonicIcon.", "Borders.", "Breakpoint.",
-			"Canvas.", "Code.", "CodeWithMe.", "CombinedDiff.", "ComboBoxButton.", "CompilationCharts.",
-			"CompletionPopup.", "ComplexPopup.", "Content.", "ContextHelp.", "CurrentMnemonic.", "Counter.",
+			"Canvas.", "CellEditor.", "Code.", "CodeWithMe.", "ColumnControlButton.", "CombinedDiff.", "ComboBoxButton.",
+			"CompilationCharts.", "CompletionPopup.", "ComplexPopup.", "Content.", "ContextHelp.", "CurrentMnemonic.", "Counter.",
 			"Debugger.", "DebuggerPopup.", "DebuggerTabs.", "DefaultTabs.", "Dialog.", "DialogWrapper.",
 			"DisclosureButton.", "DragAndDrop.",
 			"Editor.", "EditorGroupsTabs.", "EditorTabs.",
-			"FileColor.", "FlameGraph.", "Focus.",
+			"FileColor.", "FindPopup.", "FlameGraph.", "Focus.",
 			"Git.", "Github.", "GotItTooltip.", "Group.", "Gutter.", "GutterTooltip.",
 			"HeaderColor.", "HelpTooltip.", "Hg.",
-			"IconBadge.", "InformationHint.", "InplaceRefactoringPopup.",
+			"IconBadge.", "InformationHint.", "InlineBanner.", "InplaceRefactoringPopup.",
 			"Lesson.", "LineProfiler.", "Link.", "LiveIndicator.",
-			"MainMenu.", "MainToolbar.", "MemoryIndicator.", "MlModelBinding.", "MnemonicIcon.",
+			"MainMenu.", "MainToolbar.", "MainWindow.", "MemoryIndicator.", "MlModelBinding.", "MnemonicIcon.",
 			"NavBar.", "NewClass.", "NewPSD.", "Notification.", "Notifications.", "NotificationsToolwindow.",
 			"OnePixelDivider.", "OptionButton.", "Outline.",
-			"ParameterInfo.", "Plugins.", "Profiler.", "ProgressIcon.", "PsiViewer.",
-			"ReviewList.", "RunWidget.",
+			"ParameterInfo.", "PresentationAssistant.", "Plugins.", "Profiler.", "ProgressIcon.", "PsiViewer.",
+			"Resizable.", "Review.", "ReviewList.", "RunToolbar.", "RunWidget.",
 			"ScreenView.", "SearchEverywhere.", "SearchFieldWithExtension.", "SearchMatch.", "SearchOption.",
 			"SearchResults.", "SegmentedButton.", "Settings.", "SidePanel.", "Space.", "SpeedSearch.", "StateWidget.",
-			"StatusBar.",
+			"StatusBar.", "StripeToolbar.",
 			"Tag.", "TipOfTheDay.", "ToolbarComboWidget.", "ToolWindow.", "TrialWidget.",
 			"UIDesigner.", "UnattendedHostStatus.",
 			"ValidationTooltip.", "VersionControl.",
@@ -689,6 +708,9 @@ public class IntelliJTheme
 			"TabbedPane.selectedForeground"
 		) );
 
+		// Button
+		uiKeyMapping.put( "Button.minimumSize", "" ); // ignore (used in Material Theme UI Lite)
+
 		// ComboBox
 		uiKeyMapping.put( "ComboBox.background",                        "" ); // ignore
 		uiKeyMapping.put( "ComboBox.buttonBackground",                  "" ); // ignore
@@ -697,8 +719,6 @@ public class IntelliJTheme
 		uiKeyMapping.put( "ComboBox.ArrowButton.disabledIconColor",     "ComboBox.buttonDisabledArrowColor" );
 		uiKeyMapping.put( "ComboBox.ArrowButton.iconColor",             "ComboBox.buttonArrowColor" );
 		uiKeyMapping.put( "ComboBox.ArrowButton.nonEditableBackground", "ComboBox.buttonBackground" );
-		uiKeyCopying.put( "ComboBox.buttonSeparatorColor",              "Component.borderColor" );
-		uiKeyCopying.put( "ComboBox.buttonDisabledSeparatorColor",      "Component.disabledBorderColor" );
 
 		// Component
 		uiKeyMapping.put( "Component.inactiveErrorFocusColor",   "Component.error.borderColor" );
@@ -706,16 +726,16 @@ public class IntelliJTheme
 		uiKeyMapping.put( "Component.inactiveWarningFocusColor", "Component.warning.borderColor" );
 		uiKeyMapping.put( "Component.warningFocusColor",         "Component.warning.focusedBorderColor" );
 
+		// Label
+		uiKeyMapping.put( "Label.disabledForegroundColor", "" ); // ignore (used in Material Theme UI Lite)
+
 		// Link
 		uiKeyMapping.put( "Link.activeForeground", "Component.linkColor" );
 
 		// Menu
 		uiKeyMapping.put( "Menu.border",                "Menu.margin" );
 		uiKeyMapping.put( "MenuItem.border",            "MenuItem.margin" );
-		uiKeyCopying.put( "CheckBoxMenuItem.margin",    "MenuItem.margin" );
-		uiKeyCopying.put( "RadioButtonMenuItem.margin", "MenuItem.margin" );
 		uiKeyMapping.put( "PopupMenu.border",           "PopupMenu.borderInsets" );
-		uiKeyCopying.put( "MenuItem.underlineSelectionColor", "TabbedPane.underlineColor" );
 
 		// IDEA uses List.selectionBackground also for menu selection
 		uiKeyCopying.put( "Menu.selectionBackground",                "List.selectionBackground" );
@@ -723,13 +743,11 @@ public class IntelliJTheme
 		uiKeyCopying.put( "CheckBoxMenuItem.selectionBackground",    "List.selectionBackground" );
 		uiKeyCopying.put( "RadioButtonMenuItem.selectionBackground", "List.selectionBackground" );
 
-		// ProgressBar
+		// ProgressBar: IDEA uses ProgressBar.trackColor and ProgressBar.progressColor
 		uiKeyMapping.put( "ProgressBar.background",    "" ); // ignore
 		uiKeyMapping.put( "ProgressBar.foreground",    "" ); // ignore
 		uiKeyMapping.put( "ProgressBar.trackColor",    "ProgressBar.background" );
 		uiKeyMapping.put( "ProgressBar.progressColor", "ProgressBar.foreground" );
-		uiKeyCopying.put( "ProgressBar.selectionForeground", "ProgressBar.background" );
-		uiKeyCopying.put( "ProgressBar.selectionBackground", "ProgressBar.foreground" );
 
 		// ScrollBar
 		uiKeyMapping.put( "ScrollBar.trackColor", "ScrollBar.track" );
@@ -739,33 +757,29 @@ public class IntelliJTheme
 		uiKeyMapping.put( "Separator.separatorColor", "Separator.foreground" );
 
 		// Slider
+		uiKeyMapping.put( "Slider.buttonColor", "Slider.thumbColor" );
+		uiKeyMapping.put( "Slider.buttonBorderColor", "" ); // ignore
+		uiKeyMapping.put( "Slider.thumb", "" ); // ignore (used in Material Theme UI Lite)
+		uiKeyMapping.put( "Slider.track", "" ); // ignore (used in Material Theme UI Lite)
+		uiKeyMapping.put( "Slider.trackDisabled", "" ); // ignore (used in Material Theme UI Lite)
 		uiKeyMapping.put( "Slider.trackWidth", "" ); // ignore (used in Material Theme UI Lite)
-		uiKeyCopying.put( "Slider.trackValueColor", "ProgressBar.foreground" );
-		uiKeyCopying.put( "Slider.thumbColor", "ProgressBar.foreground" );
-		uiKeyCopying.put( "Slider.trackColor", "ProgressBar.background" );
-
-		// Spinner
-		uiKeyCopying.put( "Spinner.buttonSeparatorColor",         "Component.borderColor" );
-		uiKeyCopying.put( "Spinner.buttonDisabledSeparatorColor", "Component.disabledBorderColor" );
 
 		// TabbedPane
 		uiKeyMapping.put( "DefaultTabs.underlinedTabBackground", "TabbedPane.selectedBackground" );
 		uiKeyMapping.put( "DefaultTabs.underlinedTabForeground", "TabbedPane.selectedForeground" );
 		uiKeyMapping.put( "DefaultTabs.inactiveUnderlineColor",  "TabbedPane.inactiveUnderlineColor" );
+		uiKeyMapping.put( "TabbedPane.tabAreaInsets", "" ); // ignore (used in Material Theme UI Lite)
+
+		// TableHeader
+		uiKeyMapping.put( "TableHeader.cellBorder", "" ); // ignore (used in Material Theme UI Lite)
+		uiKeyMapping.put( "TableHeader.height", "" ); // ignore (used in Material Theme UI Lite)
 
 		// TitlePane
-		uiKeyCopying.put( "TitlePane.inactiveBackground",     "TitlePane.background" );
 		uiKeyMapping.put( "TitlePane.infoForeground",         "TitlePane.foreground" );
 		uiKeyMapping.put( "TitlePane.inactiveInfoForeground", "TitlePane.inactiveForeground" );
 
 		for( Map.Entry<String, String> e : uiKeyMapping.entrySet() )
 			uiKeyInverseMapping.put( e.getValue(), e.getKey() );
-
-		uiKeyCopying.put( "ToggleButton.tab.underlineColor",         "TabbedPane.underlineColor" );
-		uiKeyCopying.put( "ToggleButton.tab.disabledUnderlineColor", "TabbedPane.disabledUnderlineColor" );
-		uiKeyCopying.put( "ToggleButton.tab.selectedBackground",     "TabbedPane.selectedBackground" );
-		uiKeyCopying.put( "ToggleButton.tab.hoverBackground",        "TabbedPane.hoverColor" );
-		uiKeyCopying.put( "ToggleButton.tab.focusBackground",        "TabbedPane.focusColor" );
 
 		checkboxKeyMapping.put( "Checkbox.Background.Default",  "CheckBox.icon.background" );
 		checkboxKeyMapping.put( "Checkbox.Background.Disabled", "CheckBox.icon.disabledBackground" );
@@ -819,17 +833,15 @@ public class IntelliJTheme
 		}
 
 		@Override
-		void applyAdditionalDefaults( UIDefaults defaults ) {
-			theme.applyProperties( defaults );
+		void applyAdditionalProperties( Properties properties ) {
+			theme.applyProperties( properties );
 		}
 
 		@Override
 		protected ArrayList<Class<?>> getLafClassesForDefaultsLoading() {
-			ArrayList<Class<?>> lafClasses = new ArrayList<>();
-			lafClasses.add( FlatLaf.class );
-			lafClasses.add( theme.dark ? FlatDarkLaf.class : FlatLightLaf.class );
-			lafClasses.add( theme.dark ? FlatDarculaLaf.class : FlatIntelliJLaf.class );
-			lafClasses.add( ThemeLaf.class );
+			ArrayList<Class<?>> lafClasses = UIDefaultsLoader.getLafClassesForDefaultsLoading( getClass() );
+			lafClasses.add( 1, theme.dark ? FlatDarkLaf.class : FlatLightLaf.class );
+			lafClasses.add( 2, theme.dark ? FlatDarculaLaf.class : FlatIntelliJLaf.class );
 			return lafClasses;
 		}
 	}